+++ App.tsx
@@ -2,8 +2,8 @@ import { config } from "@gluestack-ui/config"
 import { GluestackUIProvider, Center, StatusBar, SafeAreaView, ScrollView } from "@gluestack-ui/themed"
 import { QueryClient, QueryClientProvider } from "@tanstack/react-query"
 import { Dimensions } from "react-native"
-import MySwiper from "./components/MySwiper"
-import WebtoonList from "./components/WebtoonList"
+import MySwiper from "./components/Header/MySwiper"
+import WebtoonList from "./components/List/WebtoonList"
 
 const queryClient = new QueryClient();
