+++ components/List/WebtoonList.tsx
@@ -1,8 +1,8 @@
 import  React  from 'react'
 import { useQuery } from "@tanstack/react-query"
-import Swiper from "react-native-swiper"
-import { Webtoon } from "../types"
-import LargeCard from "./Card/LargeCard"
+import { Webtoon } from "../../types"
+import { VStack } from "@gluestack-ui/themed"
+import MediumCard from "../Card/MediumCard"
 
 interface WebtoonResponse {
     webtoons: Webtoon[];
@@ -13,17 +13,17 @@ const fetchWebtoons = async() => { // 비동기 함수, promise 반환
     return response.json();
 }
 
-export default function MySwiper() {
+export default function WebtoonList() {
     const { data } = useQuery<WebtoonResponse> ({
         queryKey: ['https://korea-webtoon-api.herokuapp.com'],
         queryFn: fetchWebtoons, // 쿼리를 실행하는 비동기 함수
     });
 
   return (
-    <Swiper showsButtons={false}>
+    <VStack padding={10} bg='$white' gap={10}> 
         { data ? data.webtoons.map((webtoon) => (
-        <LargeCard key={webtoon.webtoonId} webtoon={webtoon} />
+            <MediumCard key={webtoon.webtoonId} webtoon={webtoon} />
         )) : []}
-    </Swiper>
+    </VStack>
   );
 }
